pyproject.toml
[project]
name = "StatsBot_MKWorld"
version = "0.1.0"
description = ""
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "dotenv>=0.9.9",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0",
    "pytest>=8.0",
    "ruff>=0.8",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
